updates of second review part
