+++ apps/one-drive/.env.test
@@ -4,13 +4,12 @@ MICROSOFT_REDIRECT_URI="http://localhost:4000/auth"
 MICROSOFT_INSTALL_URL="https://login.microsoftonline.com/organizations/adminconsent"
 MICROSOFT_API_URL="https://graph.microsoft.com/v1.0"
 MICROSOFT_AUTH_API_URL="https://login.microsoftonline.com"
-MICROSOFT_CHUNK_SIZE="15"
 ELBA_SOURCE_ID="0442f541-45d2-487a-9e4b-de03ce4c559e"
 ELBA_API_KEY="abc123dfg456"
 ELBA_API_BASE_URL="http://127.0.0.1:3522/api/rest"
 ELBA_REDIRECT_URL="https://foo.bar"
 ELBA_WEBHOOK_SECRET="some-elbe-webhook-secret"
-USERS_SYNC_MAX_RETRY="3"
 ENCRYPTION_KEY=801e3f2daf7a648ee7630c770f629f5b3b5d42c017109c1e2fd7e0b1570bc4cc
 WEBHOOK_URL="https://webhook.url"
 SUBSCRIBE_EXPIRATION_DAYS=2
+VERCEL_ENV="development"
